+++ src/main/java/pl/plajer/villagedefense/ConfigPreferences.java
@@ -73,7 +73,7 @@ public class ConfigPreferences {
     BUNGEE_ENABLED("BungeeActivated", false), HOLOGRAMS_ENABLED("HologramsActivated", false),
     UPGRADES_ENABLED("UpgradesActivated", false), ENABLE_SHORT_COMMANDS("Enable-Short-Commands", false),
     DISABLE_SEPARATE_CHAT("Disable-Separate-Chat", false), DISABLE_PARTIES("Disable-Parties", true),
-    RESPAWN_AFTER_WAVE("Respawn-After-Wave", true);
+    RESPAWN_AFTER_WAVE("Respawn-After-Wave", true), INGAME_JOIN_RESPAWN("InGame-Join-Respawn", true);
 
     private String path;
     private boolean def;